Move routes for 'internal' API to `/api/v1`
