+++ config/routes.rb
@@ -1,4 +1,10 @@
 Rails.application.routes.draw do
   get '/ping', to: 'status#ping', format: :json
   get '/health', to: 'status#health', format: :json
+
+  namespace :api do
+    namespace :v1 do
+      resources :moves, only: :index
+    end
+  end
 end
